game-backend-ai/pyproject.toml
[project]
name = "game-backend-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "absl-py==2.2.2",
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "astunparse==1.6.3",
    "asyncpg==0.30.0",
    "certifi==2025.4.26",
    "charset-normalizer==3.4.2",
    "click==8.2.0",
    "fastapi==0.115.12",
    "flatbuffers==25.2.10",
    "gast==0.6.0",
    "google-pasta==0.2.0",
    "grpcio==1.71.0",
    "h11==0.16.0",
    "h5py==3.13.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "idna==3.10",
    "iniconfig==2.1.0",
    "keras==3.9.2",
    "libclang==18.1.1",
    "markdown-it-py==3.0.0",
    "markdown==3.8",
    "numpy==2.1.3",
    "pydantic==2.11.4",
    "pytest-asyncio==1.1.0",
    "sqlalchemy==2.0.41",
    "uvicorn==0.34.2",
    "markupsafe==3.0.2",
    "mdurl==0.1.2",
    "ml-dtypes==0.5.1",
    "namex==0.0.9",
    "opt-einsum==3.4.0",
    "optree==0.15.0",
    "packaging==25.0",
    "pluggy==1.6.0",
    "protobuf==5.29.4",
    "pydantic-core==2.33.2",
    "pygments==2.19.1",
    "requests==2.32.3",
    "rich==14.0.0",
    "setuptools==80.8.0",
    "six==1.17.0",
    "sniffio==1.3.1",
    "starlette==0.46.2",
    "tensorboard==2.19.0",
    "tensorboard-data-server==0.7.2",
    "termcolor==3.1.0",
    "typing-extensions==4.13.2",
    "typing-inspection==0.4.0",
    "urllib3==2.4.0",
    "werkzeug==3.1.3",
    "wheel==0.45.1",
    "wrapt==1.17.2",
    "loguru==0.7.3",
    "python-dotenv==1.1.0",
    "fastapi-mcp==0.3.7",
    "aiokafka==0.12.0",
    "apscheduler==3.11.0",
    "qdrant-client==1.15.0",
    "langchain==0.3.27",
    "langchain-community==0.3.27",
    "redis==6.2.0",
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.12.14",
    "aiosignal==1.4.0",
    "async-timeout==5.0.1",
    "attrs==25.3.0",
    "dataclasses-json==0.6.7",
    "frozenlist==1.7.0",
    "h2==4.2.0",
    "hpack==4.1.0",
    "httpx-sse==0.4.1",
    "hyperframe==6.1.0",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "jsonschema==4.25.0",
    "jsonschema-specifications==2025.4.1",
    "langchain-core==0.3.72",
    "langchain-text-splitters==0.3.9",
    "langsmith==0.4.8",
    "marshmallow==3.26.1",
    "mcp==1.12.0",
    "multidict==6.6.3",
    "mypy-extensions==1.1.0",
    "orjson==3.11.1",
    "portalocker==3.2.0",
    "propcache==0.3.2",
    "pydantic-settings==2.10.1",
    "python-multipart==0.0.20",
    "pyyaml==6.0.2",
    "referencing==0.36.2",
    "requests-toolbelt==1.0.0",
    "rpds-py==0.26.0",
    "shellingham==1.5.4",
    "sse-starlette==2.4.1",
    "tenacity==9.1.2",
    "tomli==2.2.1",
    "tqdm==4.67.1",
    "typer==0.16.0",
    "typing-inspect==0.9.0",
    "tzlocal==5.3.1",
    "yarl==1.20.1",
    "zstandard==0.23.0",
    "sentence-transformers>=5.0.0",
    "pandas>=2.3.1",
    "pytest==8.4.1",
    "langchain-ollama>=0.3.6",
]
